Start implementing our playback engine and tinkering with Gst.
Without a doubt, the creation of a new source is probably the wrong way of doing this, and I should see about using buffers in the future, but baby steps.

Changed track items in the FlowBox to use our TrackItem class, which enables it to store the respective track information, which we use to pass along to the playback engine. Have KotoPlayerBar be in our Koto namespace so it can actually talk to our playback engine.