Start implementing playlist and audio seeking.
The plan is to treat every album / collection of tracks as a pseudo-playlist to share code with user-generated playlists. Changed Playback Engine current URI implementation to be a current_track instead and leverage our KotoTrack so we can more easily track its position in said playlist.

Started implementation of seeking in PlayerBar progressbar and ensure we update range and current position in progressbar on duration change.

Additionally ensure we ignore video tracks otherwise we get random windows with oggs that are frankly useless in our case.