Drop our hackish playing attempt and use our bus handler to check when async is done.
Basically check when we're done with our state change request, then start playback if we're requesting it. Additionally fixed inaccurate position being reported, since it was checking playbin and not our audio sink (we now have a dedicated audio sink and get the pad when needed.

Ensure we reset the progressbar state and redraw it.